\name{pensim-package}
\alias{pensim-package}
\alias{pensim}
\docType{package}
\title{
Functions and data for simulation of high-dimensional data and
parallelized repeated penalized regression
}
\description{
Simulation of continuous, correlated high-dimensional data with
time-to-event or binary response, and parallelized functions Lasso,
Ridge, and Elastic Net penalized regression.
}
\details{
\tabular{ll}{
Package: \tab pensim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-20\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
This package contains a function for simulation of correlated
high-dimensional data with binary or time-to-event response, and
functions for parallelized repeated optimizations of Lasso, Ridge, and
Elastic Net penalized regression penalties.
}
\author{
Levi Waldron

Maintainer: Levi Waldron <lwaldron.research@gmail.com>
}
\references{
Waldron L., Pintilie M., Huttenhower C.*, and Jurisica I.*   Optimized
application of penalized regression methods to diverse genomic
data. (2010). Under review.  (*equal contribution)
}
\keyword{ package }
\keyword{ survival }
\keyword{ datagen }
\keyword{ regression }
\keyword{ multivariate }

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
